+++ sqlparser/test/analysis/types2/misc_cases_test.dart
@@ -30,6 +30,10 @@ const Map<String, ResolvedType> _types = {
   'SELECT (3 * 4) = ?': ResolvedType(type: BasicType.int),
   'SELECT (3 / 4) = ?': ResolvedType(type: BasicType.int),
   'SELECT CURRENT_TIMESTAMP = ?': ResolvedType(type: BasicType.text),
+  'INSERT INTO demo DEFAULT VALUES ON CONFLICT (id) WHERE ? DO NOTHING':
+      ResolvedType.bool(),
+  'INSERT INTO demo DEFAULT VALUES ON CONFLICT DO UPDATE SET id = id WHERE ?':
+      ResolvedType.bool(),
 };
 
 SqlEngine _spawnEngine() {